+++ include/grad/transformer/metal_backend.h
@@ -1,5 +1,7 @@
 #pragma once
 
+#include <cstddef>
+
 // Metal (MPS) matmul backend, Apple Silicon only.
 //
 // Design: tensors never move. They stay CPU-owned; unified memory means the
@@ -42,13 +44,13 @@ bool fp16_active();
 // submission may have partially written C, so it throws std::runtime_error
 // rather than returning false.
 bool sgemm(const float* A, const float* B, float* C,
-           int M, int N, int K, bool transA, bool transB,
+           size_t M, size_t N, size_t K, bool transA, bool transB,
            float alpha, float beta);
 #else
 inline bool available() { return false; }
 inline bool fp16_active() { return false; }
 inline bool sgemm(const float*, const float*, float*,
-                  int, int, int, bool, bool, float, float) { return false; }
+                  size_t, size_t, size_t, bool, bool, float, float) { return false; }
 #endif
 
 }  // namespace grad::metal
